+++ Controls/_display/GridFooterCell.ts
@@ -1,65 +1,38 @@
 import { TemplateFunction } from 'UI/Base';
-import { mixin } from 'Types/util';
-import { OptionsToPropertyMixin } from 'Types/entity';
-import GridFooter from './GridFooter';
+import GridFooterRow from './GridFooterRow';
+import GridCell, {IOptions as IGridCellOptions} from './GridCell';
 
-export interface IOptions<T> {
-    owner: GridFooter<T>;
+export interface IOptions<T> extends IGridCellOptions<T> {
+    owner: GridFooterRow<T>;
     template?: TemplateFunction;
-    // ToDo | Временная опция для обеспечения вывода общего шаблона строки результатов.
-    //      | При разработке мультизаговков colspan будет сделан единообразно и для футера.
-    colspan?: boolean;
 }
 
 const DEFAULT_CELL_TEMPLATE = 'Controls/gridNew:FooterContent';
 
-export default class GridFooterCell<T> extends mixin<OptionsToPropertyMixin>(OptionsToPropertyMixin) {
-    protected _$owner: GridFooter<T>;
+export default class GridFooterCell<T> extends GridCell<T, GridFooterRow<T>> {
+    protected _$owner: GridFooterRow<T>;
     protected _$template: TemplateFunction;
-    protected _$colspan: boolean;
 
     constructor(options?: IOptions<T>) {
-        super();
-        OptionsToPropertyMixin.call(this, options);
-    }
-
-    getCellIndex(): number {
-        return this._$owner.getCellIndex(this);
-    }
-
-    isFirstColumn(): boolean {
-        return this.getCellIndex() === 0;
-    }
-
-    isLastColumn(): boolean {
-        return this.getCellIndex() === this._$owner.getCellsCount() - 1;
-    }
-
-    isMultiSelectColumn(): boolean {
-        return this._$owner.getMultiSelectVisibility() !== 'hidden' && this.isFirstColumn();
+        super(options);
     }
 
     getWrapperClasses(theme: string, style: string = 'default'): string {
         let wrapperClasses = `controls-Grid__footer-cell`;
-        const leftPadding = this._$owner.getLeftPadding();
         const isMultiSelectColumn = this.isMultiSelectColumn();
-
         wrapperClasses += ` controls-BaseControl__footer-content_theme-${theme}`;
 
         if (isMultiSelectColumn) {
             wrapperClasses += ` controls-ListView__footer__paddingLeft_withCheckboxes_theme-${theme}`;
         } else {
-            wrapperClasses += ` controls-ListView__footer__paddingLeft_${leftPadding}_theme-${theme}`;
+            wrapperClasses += ` controls-ListView__footer__paddingLeft_${this._$owner.getLeftPadding()}_theme-${theme}`;
         }
 
         return wrapperClasses;
     }
 
     getWrapperStyles(): string {
-        if (this._$colspan) {
-            return `grid-column: 1 / ${this._$owner.getColumnsCount() + 1}`;
-        }
-        return '';
+        return super.getWrapperStyles();
     }
 
     getContentClasses(theme: string): string {
