+++ Controls/_gridNew/Render/grid/GridView.wml
@@ -50,9 +50,9 @@
     </ws:if>
     <!-- Footer -->
     <ws:if data="{{ _listModel.getFooter() }}">
-        <ws:partial template="wml!Controls/_gridNew/Render/grid/Header"
+        <ws:partial template="{{ _listModel.getFooter().getTemplate() }}"
                     attr:name="footer"
-                    header="{{ _listModel.getFooter() }}"
+                    item="{{ _listModel.getFooter() }}"
                     backgroundStyle="{{ _options.backgroundStyle || _options.style }}"/>
     </ws:if>
 </div>
