+++ Controls/_gridNew/Render/grid/Item.wml
@@ -1,10 +1,10 @@
-<div attr:class="{{ item.getItemClasses(highlightOnHover, theme, style, cursor, clickable) }}">
+<div class="{{ item.getItemClasses(highlightOnHover, theme, style, cursor, clickable) }}">
    <ws:for data="gridColumn in item.getColumns()">
-      <div attr:class="{{ gridColumn.getWrapperClasses(theme, backgroundColorStyle, style, highlightOnHover) }}"
-           attr:style="{{ gridColumn.getWrapperStyles() }}">
-         <div if="{{ gridColumn.shouldDisplayMarker(marker) }}" attr:class="{{ gridColumn.getMarkerClasses(theme, style, markerClassName, itemPadding, markerPosition) }}"></div>
-         <div attr:class="{{ gridColumn.getContentClasses(theme, backgroundColorStyle, cursor, highlightOnHover) }}"
-              attr:style="{{ gridColumn.getContentStyles() }}">
+      <div class="{{ gridColumn.getWrapperClasses(theme, backgroundColorStyle, style, highlightOnHover) }}"
+           style="{{ gridColumn.getWrapperStyles() }}">
+         <div if="{{ gridColumn.shouldDisplayMarker(marker) }}" class="{{ gridColumn.getMarkerClasses(theme, style, markerClassName, itemPadding, markerPosition) }}"></div>
+         <div class="{{ gridColumn.getContentClasses(theme, backgroundColorStyle, cursor, highlightOnHover) }}"
+              style="{{ gridColumn.getContentStyles() }}">
             <ws:partial template="{{ gridColumn.getTemplate(multiSelectTpl) }}"
                         attr:key="{{ gridColumn.getInstanceId() }}"
                         gridColumn="{{ gridColumn }}"
